refactor: folder restructure
